requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
bcrypt
flask-cors
pandas
openpyxl
python-dotenv
